This commit implements some corrections in the code to make it more readable and also the removal of some comments.
