+++ src/sass/_device.scss
@@ -62,7 +62,7 @@
   z-index: 450;
   top: 13px;
   cursor: pointer;
-  //font-weight: bold;
+
   .mapView{
     background-color: white;
     color: black;
@@ -85,8 +85,6 @@
     padding: 5px;
   }
 }
-
-
 
 .deviceMapCanvas
 {
@@ -124,7 +122,6 @@
   background: rgba(255, 255, 255, 0.98);
   right: 0;
   display: inline-block;
-  //top: 185px;
   top:0;
   width: 400px;
   box-shadow: 10px 0px 50px #000000;
@@ -147,7 +144,6 @@
   width: 100%;
   height: 100%;
   float: right;
-  //border-left: 1px solid #BDBDBD;
   margin: 0;
 
   .main-title{
@@ -199,13 +195,8 @@
 
   .deviceCanvas{
     width: 100% !important;
-    // overflow-y: auto;
-    //height: 100%;
-    //overflow-y: scroll;
-     //height: calc(100% - 81px);
 
     .background-info{
-      //position: absolute;
       float: left;
       width: 100%;
       bottom: 0;
@@ -225,7 +216,6 @@
       .lst-entry-title{
         padding: 0 10px;
         height: 60px;
-        //position: relative;
         transition: linear 100ms;
         border-bottom: 1px solid #e2e2e2;
         cursor:pointer;
@@ -233,7 +223,6 @@
         {
           width: 58px;
           padding: 10px;
-          //height: 100%;
           margin: 0 10px;
           img
           {
@@ -256,7 +245,6 @@
           font-weight: normal;
         }
         .label{
-          //position: relative;
           font-family: 'Open Sans', sans-serif;
           color: #4F4F4F;
           font-weight: bold;
@@ -275,25 +263,13 @@
       }
     }
   }
-
-
-
 }
-
-
-
-
-
-
-
-
 
 .div-btn-side-painel{
   position: absolute;
   right: 385px;
   top: 30px;
   z-index: 100110;
-  // transition: all 1s;
 
   &.push-left{
     right: 386px;
